refactor(tools): centralize document locale
